remove pytorch_ie.core, pytorch_ie.utils.hydra, pytorch_ie.auto.AutoModel and pytorch_ie.auto.AutoTaskModule, but import respective modules from pie_core
