+++ src/pytorch_ie/__init__.py
@@ -1,5 +1,5 @@
 # flake8: noqa
 
-from pytorch_ie.auto import AutoModel, AutoPipeline, AutoTaskModule
+from pytorch_ie.auto import AutoPipeline
 from pytorch_ie.core import *
 from pytorch_ie.pipeline import Pipeline
