+++ src/pytorch_ie/core/__init__.py
@@ -1,7 +1,9 @@
-from .document import Annotation, AnnotationLayer, Document, annotation_field
-from .metric import DocumentMetric
-from .model import PyTorchIEModel
-from .module_mixins import (
+# from pie_core import auto, document, metric, model, module_mixins, statistic, taskmodule
+from pie_core.auto import AutoModel, AutoTaskModule
+from pie_core.document import Annotation, AnnotationLayer, Document, annotation_field
+from pie_core.metric import DocumentMetric
+from pie_core.model import PyTorchIEModel
+from pie_core.module_mixins import (
     EnterDatasetDictMixin,
     EnterDatasetMixin,
     ExitDatasetDictMixin,
@@ -9,8 +11,8 @@ from .module_mixins import (
     PreparableMixin,
     WithDocumentTypeMixin,
 )
-from .statistic import DocumentStatistic
-from .taskmodule import TaskEncoding, TaskModule
+from pie_core.statistic import DocumentStatistic
+from pie_core.taskmodule import TaskEncoding, TaskModule
 
 # backwards compatibility
 AnnotationList = AnnotationLayer
