+++ tests/taskmodules/test_transformer_re_text_classification.py
@@ -256,7 +256,7 @@ def test_encode(prepared_taskmodule, documents, encode_target):
     else:
         assert not encoding.has_targets
 
-        with pytest.raises(AssertionError, match=re.escape("task encoding has no target")):
+        with pytest.raises(ValueError, match=re.escape("task encoding has no target")):
             encoding.targets
